backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-docx==1.1.2
pydantic==2.8.2
python-multipart==0.0.9
openai==1.54.0
httpx==0.27.0
python-dotenv==1.0.0
